requirements.txt
flask==2.3.3
gunicorn==22.0.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
google-generativeai==0.8.2
numpy==1.26.4